// birthday-wheel-backend/src/generated/prisma/models/UserSpinSession.ts
export type UserSpinSessionSelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetSelect<{
  id?: boolean
  userId?: boolean
  tokenId?: boolean
  hasSpun?: boolean
  spunAt?: boolean
  wonPrizeId?: boolean
  createdAt?: boolean
  user?: boolean | Prisma.UserDefaultArgs<ExtArgs>
  token?: boolean | Prisma.QrTokenDefaultArgs<ExtArgs>
  prize?: boolean | Prisma.UserSpinSession$prizeArgs<ExtArgs>
}, ExtArgs["result"]["userSpinSession"]>
// ...
export type UserSpinSessionOmit<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetOmit<"id" | "userId" | "tokenId" | "hasSpun" | "spunAt" | "wonPrizeId" | "createdAt", ExtArgs["result"]["userSpinSession"]>
export type UserSpinSessionInclude<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  user?: boolean | Prisma.UserDefaultArgs<ExtArgs>
  token?: boolean | Prisma.QrTokenDefaultArgs<ExtArgs>
  prize?: boolean | Prisma.UserSpinSession$prizeArgs<ExtArgs>
}
// ...
export type $UserSpinSessionPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "UserSpinSession"
  objects: {
    user: Prisma.$UserPayload<ExtArgs>
    token: Prisma.$QrTokenPayload<ExtArgs>
    prize: Prisma.$PrizePayload<ExtArgs> | null
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    userId: string
    tokenId: string
    hasSpun: boolean
    spunAt: Date | null
    wonPrizeId: number | null
    createdAt: Date
  }, ExtArgs["result"]["userSpinSession"]>
  composites: {}
}

export type UserSpinSessionGetPayload<S extends boolean | null | undefined | UserSpinSessionDefaultArgs> = runtime.Types.Result.GetResult<Prisma.$UserSpinSessionPayload, S>
// ...
export type UserSpinSessionDefaultArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the UserSpinSession
   */
  select?: Prisma.UserSpinSessionSelect<ExtArgs> | null
  /**
   * Omit specific fields from the UserSpinSession
   */
  omit?: Prisma.UserSpinSessionOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.UserSpinSessionInclude<ExtArgs> | null
}